Change free space tile creation to be at runtime
Modifies how the "Free Space" tile is generated. Before, the tile was
pre-generated and the image was added to the other generated tiles at
runtime. This changes the behaviour to generate the "Free Space" tile
at runtime so the tile can be modified during development without having
to pre-generate a new image.

Change-Id: I4fbca6f96f02ec3df8cdfd8d36eba717d6855f1d